refactor: move web-api to authorization middlewares
