+++ app/shared/authorization/authorization.middleware.test.ts
@@ -4,11 +4,22 @@ import { teamFactory } from 'tests/factories/team.ts';
 import { userFactory } from 'tests/factories/users.ts';
 import type { Mock } from 'vitest';
 import { RequireAuthContext } from '../authentication/auth.middleware.ts';
-import { BadRequestError, EventNotFoundError, ForbiddenOperationError, NotFoundError } from '../errors.server.ts';
 import {
+  ApiKeyInvalidError,
+  ApiKeyQueryParamsDeprecatedError,
+  BadRequestError,
+  EventNotFoundError,
+  ForbiddenError,
+  ForbiddenOperationError,
+  NotFoundError,
+} from '../errors.server.ts';
+import { flags } from '../feature-flags/flags.server.ts';
+import {
+  AuthorizedApiEventContext,
   AuthorizedEventContext,
   AuthorizedTeamContext,
   requireAdmin,
+  requireAuthorizedApiEvent,
   requireAuthorizedEvent,
   requireAuthorizedTeam,
 } from './authorization.middleware.ts';
@@ -481,5 +492,154 @@ describe('middleware chain behavior', () => {
         mockNext,
       );
     }).rejects.toThrow(BadRequestError);
+  });
+});
+
+describe('requireAuthorizedApiEvent', () => {
+  describe('header-based authentication', () => {
+    it('sets event in context when API key is valid in header', async () => {
+      const event = await eventFactory({ attributes: { apiKey: 'valid-api-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test', {
+        headers: { 'X-API-Key': 'valid-api-key' },
+      });
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext);
+
+      const contextEvent = context.get(AuthorizedApiEventContext);
+      expect(contextEvent?.event.id).toBe(event.id);
+      expect(contextEvent?.event.slug).toBe('test-event');
+      expect(contextEvent?.event.apiKey).toBe('valid-api-key');
+    });
+
+    it('throws ApiKeyInvalidError when header API key does not match', async () => {
+      await eventFactory({ attributes: { apiKey: 'correct-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test', {
+        headers: { 'X-API-Key': 'wrong-key' },
+      });
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(ApiKeyInvalidError);
+    });
+
+    it('prefers header API key over query params', async () => {
+      const event = await eventFactory({ attributes: { apiKey: 'header-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test?key=query-key', {
+        headers: { 'X-API-Key': 'header-key' },
+      });
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext);
+
+      const contextEvent = context.get(AuthorizedApiEventContext);
+      expect(contextEvent?.event.id).toBe(event.id);
+    });
+  });
+
+  describe('query params authentication (backward compatibility)', () => {
+    it('sets event in context when API key is valid in query params', async () => {
+      const event = await eventFactory({ attributes: { apiKey: 'valid-api-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test?key=valid-api-key');
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext);
+
+      const contextEvent = context.get(AuthorizedApiEventContext);
+      expect(contextEvent?.event.id).toBe(event.id);
+      expect(contextEvent?.event.slug).toBe('test-event');
+      expect(contextEvent?.event.apiKey).toBe('valid-api-key');
+    });
+
+    it('validates API key from query string correctly', async () => {
+      await eventFactory({ attributes: { apiKey: 'my-secret-key-123', slug: 'my-event' } });
+      const request = new Request('https://example.com/api/v1/proposals?key=my-secret-key-123&filter=accepted');
+      const context = createMockContext();
+      const params = { event: 'my-event' };
+
+      await requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext);
+
+      const contextEvent = context.get(AuthorizedApiEventContext);
+      expect(contextEvent?.event.slug).toBe('my-event');
+    });
+
+    it('throws ApiKeyQueryParamsDeprecatedError when query params are disabled', async () => {
+      await flags.set('disableApiKeyInQueryParams', true);
+      await eventFactory({ attributes: { apiKey: 'valid-api-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test?key=valid-api-key');
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(ApiKeyQueryParamsDeprecatedError);
+
+      await flags.set('disableApiKeyInQueryParams', false);
+    });
+
+    it('allows header authentication when query params are disabled', async () => {
+      await flags.set('disableApiKeyInQueryParams', true);
+      const event = await eventFactory({ attributes: { apiKey: 'valid-api-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test', {
+        headers: { 'X-API-Key': 'valid-api-key' },
+      });
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext);
+
+      const contextEvent = context.get(AuthorizedApiEventContext);
+      expect(contextEvent?.event.id).toBe(event.id);
+
+      await flags.set('disableApiKeyInQueryParams', false);
+    });
+  });
+
+  describe('error handling', () => {
+    it('throws ForbiddenError when API key query parameter is missing', async () => {
+      const request = new Request('https://example.com/api/test');
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(ForbiddenError);
+    });
+
+    it('throws EventNotFoundError when event slug is not in params', async () => {
+      const request = new Request('https://example.com/api/test?key=some-key');
+      const context = createMockContext();
+      const params = {};
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(EventNotFoundError);
+    });
+
+    it('throws EventNotFoundError when event does not exist', async () => {
+      const request = new Request('https://example.com/api/test?key=valid-api-key');
+      const context = createMockContext();
+      const params = { event: 'non-existent-event' };
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(EventNotFoundError);
+    });
+
+    it('throws ApiKeyInvalidError when API key does not match', async () => {
+      await eventFactory({ attributes: { apiKey: 'correct-key', slug: 'test-event' } });
+      const request = new Request('https://example.com/api/test?key=wrong-key');
+      const context = createMockContext();
+      const params = { event: 'test-event' };
+
+      await expect(
+        requireAuthorizedApiEvent({ request, context, params, unstable_pattern: '' }, mockNext),
+      ).rejects.toThrow(ApiKeyInvalidError);
+    });
   });
 });
